requirements.txt
absl-py
aioice
aiortc
altair
astunparse
attrs
av
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cvzone
cycler
dnspython
filelock
Flask
flatbuffers
fonttools
fsspec
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-crc32c
google-pasta
grpcio
h5py
idna
ifaddr
itsdangerous
jax
jaxlib
Jinja2
jsonschema
jsonschema-specifications
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
mpmath
namex
narwhals
networkx
numpy
oauthlib
opencv-python
opt_einsum
optree
packaging
pandas
pillow
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pycparser
pydeck
pyee
Pygments
pylibsrtp
pyOpenSSL
pyparsing
python-dateutil
pytz
PyYAML
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
scipy
seaborn
six
smmap
streamlit
streamlit-webrtc
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
toml
torch
torchaudio
torchvision
tornado
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
watchdog
Werkzeug
wrapt
